More parsing cleanup
