--- src/lib/index.ts
@@ -1,4 +1,0 @@
-export * from "../parser/char-stream.js";
-export * from "../parser/token.js";
-export * as grammar from "./grammar.js";
-export * from "../parser/api/parse-file.js";
